src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.3
attrs==23.1.0
bellows==0.37.2
cffi==1.16.0
click==8.1.7
click-log==0.4.0
coloredlogs==15.0.1
crc==6.0.0
crccheck==1.3.0
cryptography==41.0.7
frozenlist==1.4.0
humanfriendly==10.0
idna==3.6
multidict==6.0.4
pure-pcapy3==1.0.1
pycparser==2.21
pyserial==3.5
pyserial-asyncio==0.6
typing_extensions==4.9.0
universal-silabs-flasher==0.0.15
voluptuous==0.14.1
yarl==1.9.4
zigpy==0.60.1
